+++ lib/model/social.dart
@@ -1181,7 +1181,11 @@ enum ConversationType { direct, groupSubset, groupAll }
 
 extension ConversationTypeImpl on ConversationType {
 
-  bool get isGroup => ((this == ConversationType.groupSubset) || (this == ConversationType.groupAll));
+  bool get isGroup => groupTypes.contains(this);
+  static const Set<ConversationType> groupTypes = const <ConversationType>{ ConversationType.groupSubset, ConversationType.groupAll };
+
+  bool get isDirect => directTypes.contains(this);
+  static const Set<ConversationType> directTypes = const <ConversationType>{ ConversationType.direct };
 
   static ConversationType? fromJsonString(String? value) {
     switch (value?.toLowerCase()) {